requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
_py-xgboost-mutex=2.0=gpu_0
astroid=3.2.2=pypi_0
asttokens=2.4.0=pyhd8ed1ab_0
atk-1.0=2.36.0=ha1a6a79_0
awkward=0.14.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
blosc=1.21.3=h6a678d5_0
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
c-blosc2=2.8.0=h6a678d5_0
ca-certificates=2023.12.12=h06a4308_0
cairo=1.16.0=hb05425b_5
certifi=2023.11.17=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.0.5=py39hdb19cb5_0
cryptography=41.0.3=py39hdda0065_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cuda-version=12.2=he2b69de_2
curl=8.2.1=hdbd6064_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.8=pypi_0
einops=0.6.1=pypi_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py39h06a4308_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.1=h4c34cd2_2
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fribidi=1.0.10=h7b6447c_0
gdk-pixbuf=2.42.10=h5eee18b_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
gobject-introspection=1.72.0=py39hbb6d50b_2
graphite2=1.3.14=h295c915_1
graphviz=2.50.0=h1b29801_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
gtk2=2.24.33=h73c1081_2
gts=0.7.6=hb67d8dd_3
h5py=3.9.0=py39he06866b_0
harfbuzz=4.3.0=hf52aaf7_1
hdf5=1.12.1=h2b7332f_3
icu=58.2=he6710b0_3
idna=3.4=py39h06a4308_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.25.2=pyh2140261_0
ipython=8.15.0=pyh0d859eb_0
isort=5.13.2=pypi_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=py39h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=8.3.1=pyhd8ed1ab_0
jupyter_core=4.12.0=py39hf3d152e_0
kaggle=1.5.16=pypi_0
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.2.10=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.3.141=0
libcurl=8.2.1=h251f7ec_0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_3
libgd=2.3.3=h695aa2c_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.52.0=h2d74bed_1
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
librsvg=2.54.4=h36cc946_3
libsodium=1.0.18=h36c2ea0_1
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.2.0=h7e041cc_3
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libtool=2.4.6=h6a678d5_1009
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxgboost=1.7.6=cuda120_h75debf4_6
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lightning-utilities=0.9.0=pyhd8ed1ab_0
linformer=0.2.1=pypi_0
lion-pytorch=0.1.2=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
lzo=2.10=h7b6447c_2
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.7.2=py39h06a4308_0
matplotlib-base=3.7.2=py39h1128e8f_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mccabe=0.7.0=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
nccl=2.19.4.1=h3a97aeb_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.8.4=py39hc78ab66_1
numpy=1.25.2=py39h5f9d8c6_0
numpy-base=1.25.2=py39hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
packaging=23.1=py39h06a4308_0
pandas=2.0.3=py39h1128e8f_0
pango=1.50.7=h05da053_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py39h6a678d5_1
pip=23.2.1=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
platformdirs=4.2.2=pypi_0
ply=3.11=py39h06a4308_0
poppler=0.81.0=h01f5e8b_2
poppler-data=0.4.11=h06a4308_1
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-cpuinfo=8.0.0=pyhd3eb1b0_1
py-xgboost=1.7.6=cuda120_py39h206be00_6
py-xgboost-gpu=1.7.6=py39h206be00_6
pycparser=2.21=pyhd3eb1b0_0
pygments=2.16.1=pyhd8ed1ab_0
pylint=3.2.2=pypi_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.15.7=py39h6a678d5_1
pyqt5-sip=12.11.0=py39h6a678d5_1
pysocks=1.7.1=py39h06a4308_0
pytables=3.8.0=py39hb8ae3fc_3
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.20.1=pypi_0
python-slugify=8.0.1=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=2.0.1=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py39h06a4308_0
pyzmq=25.1.0=py39h6a678d5_0
qt-main=5.15.2=h7358343_9
qt-webengine=5.15.9=h9ab4d14_7
qtwebkit=5.212=h3fafdc1_5
readline=8.2=h5eee18b_0
requests=2.31.0=py39h06a4308_0
scikit-learn=1.3.0=py39h1128e8f_0
scipy=1.11.1=py39h5f9d8c6_0
setuptools=68.0.0=py39h06a4308_0
sip=6.6.2=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
text-unidecode=1.3=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=pypi_0
tomlkit=0.12.5=pypi_0
torchaudio=2.0.2=py39_cu117
torcheval=0.0.7=pypi_0
torchmetrics=1.1.2=pyhd8ed1ab_0
torchsummary=1.5.1=pypi_0
torchtriton=2.0.0=py39
torchvision=0.15.2=py39_cu117
torchviz=0.0.2=pypi_0
tornado=6.3.2=py39h5eee18b_0
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.10.0=pyhd8ed1ab_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py39h06a4308_0
vit-pytorch=1.4.5=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
wheel=0.38.4=py39h06a4308_0
xz=5.4.2=h5eee18b_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.11.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
zlib-ng=2.0.7=h5eee18b_0
zstd=1.5.5=hc292b87_0
